split single long spec into individual specs
