+++ examples/stubbing-spying__window-fetch/cypress/integration/deferred.js
@@ -1,0 +1,10 @@
+// little utility for making Promise-returning stubs easier
+module.exports = function () {
+  const deferred = {}
+  /* global Promise */
+  deferred.promise = new Promise((resolve, reject) => {
+    deferred.resolve = resolve
+    deferred.reject = reject
+  })
+  return deferred
+}
